requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==1.8.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
Automat==20.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==4.1.0
blinker==1.4
blis==1.2.0
bottle==0.12.19
Brotli==1.0.9
catalogue==2.0.10
certifi==2024.12.14
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloud-init==24.3.1
cloudpathlib==0.20.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
commonmark==0.9.1
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
ctop==1.0.0
cycler==0.11.0
cymem==2.0.11
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.12
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.17
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.4
distro==1.9.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.6.0
filetype==1.2.0
flake8==4.0.1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.12
fsspec==2024.12.0
future==0.18.2
gast==0.5.2
Glances==3.2.4.2
google-pasta==0.2.0
graphiti-core==0.5.2
greenlet==3.1.1
grpcio==1.30.2
gyp==0.1
h11==0.14.0
h5py==3.6.0
h5py.-debian-h5py-serial==3.6.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.27.1
hyperlink==21.0.0
icdiff==2.0.4
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
influxdb==5.3.1
iotop==0.6
ipykernel==6.29.5
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaptan==0.5.12
keras==3.6.0
keyring==23.5.0
kiwisolver==1.3.2
langchain-core==0.3.31
langchain-huggingface==0.1.2
langchain-openai==0.3.2
langcodes==3.5.0
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.1
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libtmux==0.10.1
livereload==2.6.3
llama-cloud==0.1.10
llama-index==0.12.14
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
lxml==4.8.0
lz4==3.1.3+dfsg
marisa-trie==1.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.26.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==3.1.0
mkdocs==1.1.2
ml-dtypes==0.5.0
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.21.5
nvidia-ml-py==12.555.43
oauthlib==3.2.0
olefile==0.46
openai==1.60.1
opt-einsum==3.3.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pipx==1.0.0
platformdirs==2.5.1
ply==3.11
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.21.12
psutil==5.9.0
psycopg==3.2.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==5.2.0
PyPDF2==3.0.1
pyrsistent==0.18.1
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.12
pystache==0.6.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-magic==0.4.24
pythran==0.10.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==0.23.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.3
sentence-transformers==3.4.0
service-identity==18.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sos==4.7.2
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
ssh-import-id==5.11
striprtf==0.0.26
sympy==1.12
tenacity==9.0.0
tensorboard==2.18.0
tensorflow==2.18.0
termcolor==1.1.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.1.0
tiktoken==0.8.0
tmuxp==1.9.2
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
Twisted==22.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==2.3.0
userpath==1.8.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.2
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.23.0
